delegated message input into a separate component 'InputBox' to reduce the horrible lag that builds up in the 'Messages' component. There is less lag but its still present
